HU-025.3 Visualizar listado de movimientos de Kardex

+++ src/main/java/com/coagronet/articuloPedido/repositories/ArticuloPedidoRepository.java
@@ -1,5 +1,6 @@
 package com.coagronet.articuloPedido.repositories;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,7 +20,7 @@ public interface ArticuloPedidoRepository extends JpaRepository<ArticuloPedido, 
 
 		Long getPresentacionId();
 
-		Double getCantidad();
+		BigDecimal getCantidad();
 
 	}
